refactor: remove get decl layer
